+++ src/Entity/Product.php
@@ -277,7 +277,7 @@ class Product
     /**
      * @return int
      */
-    public function getNbrSignatures(): int
+    public function getNbrSignatures(): float
     {
         $sonde = $this->getSonde();
         $colonizer = $this->getColonizer() * 200;
